# api/e2etests/api/openapi.yaml
openapi: 3.0.3
info:
  description: |
    This is a prototype of Talknow Project's front/back API.
  title: Talknow API
  version: 0.0.1
servers:
- url: http://localhost:8080/
security:
- CookieAuth: []
...
paths:
...
  /profile/contact:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ContactInfoDTO'
                type: array
          description: Les contacts ont bien été récupérés.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
          description: Error
      security:
      - CookieAuth: []
      summary: Obtenir la liste de tous ses contacts
      tags:
      - contact
      x-accepts: application/json
    post:
      description: "Si l'utilisateur souhaite discuter avec un contact qui n'existe\
        \ pas encore dans sa base de données, il peut l'ajouter."
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AjoutContactDTO'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ContactInfoDTO'
                type: array
          description: Contact créé avec succès et on renvoie tous les champs de contacts
            remplis avec un avatar par défaut et un login qu'on déduit de l'adresse
            e-mail.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
          description: Error
      security:
      - CookieAuth: []
      summary: Permet à l'utilisateur d'ajouter un contact.
      tags:
      - contact
      x-content-type: application/json
      x-accepts: application/json
  /profile/contact/{contactid}:
    delete:
      description: "Si l'utilisateur ne souhaite plus avoir un contact donné, il dispose\
        \ d'une option pour le supprimer de sa base de données."
      parameters:
      - description: Contact's id
        explode: false
        in: path
        name: contactid
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ContactInfoDTO'
                type: array
          description: "Le contact a été supprimé avec succès, on retourne la liste\
            \ des contacts mise à jour"
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
          description: Error
      security:
      - CookieAuth: []
      summary: Permet à l'utilisateur de supprimer un de ses contacts.
      tags:
      - contact
      x-accepts: application/json
    post:
      description: "Si l'utilisateur souhaite discuter avec un utilisateur pour la\
        \ première fois, en appuyant sur le bouton une conversation se crée au niveau\
        \ du home."
      parameters:
      - description: Contact's id
        explode: false
        in: path
        name: contactid
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: Conversation créée avec succès.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
          description: Error
      security:
      - CookieAuth: []
      summary: Permet à l'utilisateur de créer une conversation avec un contact.
      tags:
      - contact
      x-accepts: application/json
    put:
      description: "Si l'utilisateur souhaite renommé un de ses contacts, il peut\
        \ effectué ceci, ainsi que modifier l'avatar du contact."
      parameters:
      - description: Contact's id
        explode: false
        in: path
        name: contactid
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInfoDTO'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ContactInfoDTO'
                type: array
          description: "Le contact a été modifié avec succès, on retourne la liste\
            \ des contacts mise à jour"
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
          description: Error
      security:
      - CookieAuth: []
      summary: Permet à l'utilisateur de modifier des informations sur le contact
        souhaité.
      tags:
      - contact
      x-content-type: application/json
      x-accepts: application/json
...
components:
  schemas:
...
    UpdateInfoDTO:
      example:
        type: type
        value: value
      properties:
        type:
          type: string
        value:
          type: string
      type: object
...
    ContactInfoDTO:
      example:
        firstname: firstname
        avatar: avatar
        id: 0
        pseudo: pseudo
        email: email
        lastname: lastname
      properties:
        firstname:
          type: string
        lastname:
          type: string
        email:
          type: string
        pseudo:
          type: string
        avatar:
          type: string
        id:
          format: int64
          type: integer
      type: object
    AjoutContactDTO:
      example:
        email: email
      properties:
        email:
          type: string
      type: object
...
    ErrorDTO:
      properties:
        timestamp:
          type: string
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
      type: object
  securitySchemes:
    CookieAuth:
      description: Use a session cookie to authenticate (see /login).
      in: cookie
      name: JSESSIONID
      type: apiKey
...