# server/src/main/java/fr/mightycode/cpoo/server/dto/AjoutContactDTO.java
package fr.mightycode.cpoo.server.dto;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.Setter;

@AllArgsConstructor
@Getter
@Setter
public class AjoutContactDTO {
  String email;
  public AjoutContactDTO(){

  }
}